requirements.txt
selenium==4.40.0
pytest==9.0.2
allure-pytest==2.15.3